Create modular transformer API
